Hw5

+++ bot/src/main/java/edu/java/bot/dto/bot/PostRequest.java
@@ -6,5 +6,5 @@ import java.util.List;
 public record PostRequest(@JsonProperty("id") Integer id,
                           @JsonProperty("url") String url,
                           @JsonProperty("description") String description,
-                          @JsonProperty("tgChatIds") List<Integer> tgChatIds) {
+                          @JsonProperty("tgChatIds") List<Long> tgChatIds) {
 }
