+++ bot/src/main/java/edu/java/bot/dto/scrapper/DeleteResponse.java
@@ -3,6 +3,6 @@ package edu.java.bot.dto.scrapper;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import java.net.URI;
 
-public record DeleteResponse(@JsonProperty("id") Integer id,
+public record DeleteResponse(@JsonProperty("id") Long id,
                              @JsonProperty("url") URI url) {
 }
