+++ bot/src/main/java/edu/java/bot/dto/scrapper/PostResponse.java
@@ -3,6 +3,6 @@ package edu.java.bot.dto.scrapper;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import java.net.URI;
 
-public record PostResponse(@JsonProperty("id") Integer id,
+public record PostResponse(@JsonProperty("id") Long id,
                            @JsonProperty("url") URI url) {
 }
